Rename: 대소문자 변경 위해 폴더명 test로 변경
