+++ app/(route)/(auth)/signup/_components/sign-up-form.tsx
@@ -10,7 +10,7 @@ import { postSignup } from '@/_apis/user';
 import { useModal } from '@/_hooks/useModal';
 
 import Button from '@/_components/button';
-import Input from '@/_components/Input';
+import Input from '@/_components/InputTest';
 import Modal from '@/_components/modal';
 
 interface InputField {
